# be/apidocs/swagger.yaml
basePath: /api/
definitions:
...
  controller.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controller.LostPasswordInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  controller.ModifyUserInput:
    properties:
      barbershopsId:
        items:
          type: string
        type: array
      email:
        type: string
    type: object
  controller.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  entity.Appointment:
    properties:
      barbershopID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      start:
        type: string
      userID:
        type: string
    type: object
  entity.BarberShop:
    properties:
      address:
        type: string
      appointments:
        items:
          $ref: '#/definitions/entity.Appointment'
        type: array
      employees:
        type: integer
      id:
        type: string
      imageLink:
        type: string
      location:
        $ref: '#/definitions/entity.Location'
      name:
        type: string
      phone:
        type: string
      rating:
        type: number
      reviews:
        items:
          $ref: '#/definitions/entity.Review'
        type: array
    type: object
...
  entity.Location:
    properties:
      coordinates:
        items:
          type: number
        type: array
      type:
        type: string
    type: object
  entity.Review:
    properties:
      content:
        type: string
      createdAt:
        type: string
      rating:
        type: integer
      reported:
        type: boolean
    type: object
...
  entity.User:
    properties:
      currentAppointment:
        $ref: '#/definitions/entity.Appointment'
      email:
        type: string
      id:
        type: string
      ownedShops:
        items:
          $ref: '#/definitions/entity.BarberShop'
        type: array
      type:
        type: string
      username:
        type: string
    type: object
host: localhost:7000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gin Swagger Example API
  version: "1.0"
paths:
...
  /admin/user:
    get:
      consumes:
      - application/json
      description: Get a list of users filtered by email
      parameters:
      - description: Email filter
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/entity.User'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Show list of users
      tags:
      - User
  /admin/user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user by ID
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/entity.User'
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show user information by ID
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Modify a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Modify user input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.ModifyUserInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Modify user by ID
      tags:
      - User
...
  /user:
    post:
      consumes:
      - application/json
      description: Create new user with the provided email and password
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Create new user
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with the provided email and password and returns
        a JWT token for subsequent authenticated requests
      parameters:
      - description: User email and password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: 'token" : "token"}'
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Logs in a user and returns a JWT token
      tags:
      - Auth
  /user/lostpassword:
    post:
      consumes:
      - application/json
      description: Generate a password reset token and send it to the user's email
        address
      parameters:
      - description: Lost password input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.LostPasswordInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request password reset
      tags:
      - User
  /user/resetpassword/{resettoken}:
    post:
      consumes:
      - application/json
      description: Reset password for a user given a reset token
      parameters:
      - description: Reset token
        in: path
        name: reset_token
        required: true
        type: string
      - description: New password
        in: body
        name: newpassword
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset user password
      tags:
      - User
  /user/self:
    get:
      consumes:
      - application/json
      description: Get user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/entity.User'
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show user information by ID
      tags:
      - User
...
schemes:
- http
swagger: "2.0"